making progress with cmake

+++ C++/Referencias/char_arrays/char_arrays.cpp
@@ -4,7 +4,7 @@ using namespace std;
 int main(){
   char texto[] = "hola";
 
-  for(int i = 0; i< sizeof(texto); i++){
+  for(size_t i = 0; i< sizeof(texto); i++){
     cout << i <<": " << texto[i] << endl;
   }
